requirements.txt.txt
langchain 
chromadb 
pypdf 
sentence_transformers 
openai
python-dotenv
flask